Add the foundation of self-hosted authentication  (#113)

+++ src/helpers/profile/deleteProfile.ts
@@ -3,6 +3,7 @@ import toast from 'react-hot-toast'
 import { DELETE_PROFILE } from '../../graphql/profile'
 import { Profile } from '../../models/profile'
 import { firebaseApp } from '../../pages'
+import { firebaseEnabled } from '../getFirebaseConfig'
 import { getHasuraClient } from '../getHasuraClient'
 import { handleError } from '../handleError'
 import { logout } from '../logout'
@@ -19,8 +20,10 @@ export const deleteProfile = async (profile: Profile) => {
       await client.request(stringifyQuery(DELETE_PROFILE), {
         authId: profile.authId,
       })
-      const auth = getAuth(firebaseApp)
-      await auth.currentUser?.delete()
+      if (firebaseEnabled) {
+        const auth = getAuth(firebaseApp)
+        await auth.currentUser?.delete()
+      }
       toast.success('Your account has been deleted!')
       await logout(false)
     }
